requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
dataclasses-json==0.6.0
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
emoji==2.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
filetype==1.2.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.1
jupyter_core==5.3.1
langchain==0.0.300
langsmith==0.0.40
loguru==0.7.2
lxml==4.9.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr==2.8.6
numpy==1.25.2
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pandas-stubs==2.0.3.230814
parso==0.8.3
pdf2image==1.16.3
pdfminer==20191125
pickleshare==0.7.5
Pillow==10.0.1
pinecone-client==2.2.4
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pypdf==3.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
types-pytz==2023.3.0.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.10.16
urllib3==2.0.4
wcwidth==0.2.6
win32-setctime==1.1.0
yarl==1.9.2
streamlit==1.18.1
streamlit_lottie==1.0.0